requirements.txt
flask==3.0.2
google-generativeai==0.3.2
python-dotenv==1.0.1
requests==2.31.0
gunicorn==21.2.0 